+++ uxarray/grid/coordinates.py
@@ -106,6 +106,53 @@ def normalize_in_place(node):
     return np.array(node) / np.linalg.norm(np.array(node), ord=2)
 
 
+def grid_center_lat_lon(ds):
+    """Using scrip file variables ``grid_corner_lat`` and ``grid_corner_lon``,
+    calculates the ``grid_center_lat`` and ``grid_center_lon``.
+
+    Parameters
+    ----------
+    ds : xarray.Dataset
+        Dataset that contains ``grid_corner_lat`` and ``grid_corner_lon``
+        data variables
+
+    Returns
+    -------
+    center_lon : :class:`numpy.ndarray`
+        The calculated center longitudes of the grid box based on the corner
+        points
+    center_lat : :class:`numpy.ndarray`
+        The calculated center latitudes of the grid box based on the corner
+        points
+    """
+
+    # Calculate and create grid center lat/lon
+    scrip_corner_lon = ds['grid_corner_lon']
+    scrip_corner_lat = ds['grid_corner_lat']
+
+    # convert to radians
+    rad_corner_lon = np.deg2rad(scrip_corner_lon)
+    rad_corner_lat = np.deg2rad(scrip_corner_lat)
+
+    # get nodes per face
+    nodes_per_face = rad_corner_lat.shape[1]
+
+    # geographic center of each cell
+    x = np.sum(np.cos(rad_corner_lat) * np.cos(rad_corner_lon),
+               axis=1) / nodes_per_face
+    y = np.sum(np.cos(rad_corner_lat) * np.sin(rad_corner_lon),
+               axis=1) / nodes_per_face
+    z = np.sum(np.sin(rad_corner_lat), axis=1) / nodes_per_face
+
+    center_lon = np.rad2deg(np.arctan2(y, x))
+    center_lat = np.rad2deg(np.arctan2(z, np.sqrt(x**2 + y**2)))
+
+    # Make negative lons positive
+    center_lon[center_lon < 0] += 360
+
+    return center_lat, center_lon
+
+
 def _populate_cartesian_xyz_coord(grid):
     """A helper function that populates the xyz attribute in UXarray.Grid._ds.
     This function is called when we need to use the cartesian coordinates for
@@ -237,50 +284,3 @@ def _populate_lonlat_coord(grid):
             "long_name": "latitude of mesh nodes",
             "units": "degrees_north",
         })
-
-
-def grid_center_lat_lon(ds):
-    """Using scrip file variables ``grid_corner_lat`` and ``grid_corner_lon``,
-    calculates the ``grid_center_lat`` and ``grid_center_lon``.
-
-    Parameters
-    ----------
-    ds : xarray.Dataset
-        Dataset that contains ``grid_corner_lat`` and ``grid_corner_lon``
-        data variables
-
-    Returns
-    -------
-    center_lon : :class:`numpy.ndarray`
-        The calculated center longitudes of the grid box based on the corner
-        points
-    center_lat : :class:`numpy.ndarray`
-        The calculated center latitudes of the grid box based on the corner
-        points
-    """
-
-    # Calculate and create grid center lat/lon
-    scrip_corner_lon = ds['grid_corner_lon']
-    scrip_corner_lat = ds['grid_corner_lat']
-
-    # convert to radians
-    rad_corner_lon = np.deg2rad(scrip_corner_lon)
-    rad_corner_lat = np.deg2rad(scrip_corner_lat)
-
-    # get nodes per face
-    nodes_per_face = rad_corner_lat.shape[1]
-
-    # geographic center of each cell
-    x = np.sum(np.cos(rad_corner_lat) * np.cos(rad_corner_lon),
-               axis=1) / nodes_per_face
-    y = np.sum(np.cos(rad_corner_lat) * np.sin(rad_corner_lon),
-               axis=1) / nodes_per_face
-    z = np.sum(np.sin(rad_corner_lat), axis=1) / nodes_per_face
-
-    center_lon = np.rad2deg(np.arctan2(y, x))
-    center_lat = np.rad2deg(np.arctan2(z, np.sqrt(x**2 + y**2)))
-
-    # Make negative lons positive
-    center_lon[center_lon < 0] += 360
-
-    return center_lat, center_lon
