+++ uxarray/grid/grid.py
@@ -8,12 +8,16 @@ from uxarray.io._mpas import _read_mpas
 from uxarray.io._ugrid import _read_ugrid, _encode_ugrid
 from uxarray.io._shapefile import _read_shpfile
 from uxarray.io._scrip import _read_scrip, _encode_scrip
-from uxarray.io.utils import parse_grid_type
+from uxarray.io.utils import _parse_grid_type
 
 from uxarray.grid.area import get_all_face_area_from_coords
 from uxarray.grid.connectivity import (_build_edge_node_connectivity,
                                        _build_face_edges_connectivity,
                                        _build_nNodes_per_face)
+
+from uxarray.grid.coordinates import (_populate_lonlat_coord,
+                                      _populate_cartesian_xyz_coord)
+
 from uxarray.constants import INT_DTYPE, INT_FILL_VALUE
 
 
@@ -100,7 +104,7 @@ class Grid:
         # check if initializing from string
         # TODO: re-add gridspec initialization when implemented
         elif isinstance(input_obj, xr.Dataset):
-            self.mesh_type = parse_grid_type(input_obj)
+            self.mesh_type = _parse_grid_type(input_obj)
             self.__from_ds__(dataset=input_obj)
         else:
             raise RuntimeError("Dataset is not a valid input type.")
@@ -422,7 +426,7 @@ class Grid:
         Dimensions (``nMesh2_node``)
         """
         if "Mesh2_node_cart_x" not in self._ds:
-            self._populate_cartesian_xyz_coord()
+            _populate_cartesian_xyz_coord(self)
         return self._ds['Mesh2_node_cart_x']
 
     @property
@@ -454,7 +458,7 @@ class Grid:
         Dimensions (``nMesh2_node``)
         """
         if "Mesh2_node_cart_y" not in self._ds:
-            self._populate_cartesian_xyz_coord()
+            _populate_cartesian_xyz_coord(self)
         return self._ds['Mesh2_node_cart_y']
 
     @property
@@ -497,8 +501,6 @@ class Grid:
         if "Mesh2_node_cart_z" not in self._ds:
             self._populate_cartesian_xyz_coord()
         return self._ds['Mesh2_node_cart_z']
-
-    # connectivity properties
 
     @property
     def Mesh2_face_nodes(self):
