+++ uxarray/io/_scrip.py
@@ -1,7 +1,7 @@
 import xarray as xr
 import numpy as np
 
-from uxarray.grid.utils import _replace_fill_values
+from uxarray.grid.connectivity import _replace_fill_values
 from uxarray.grid.coordinates import grid_center_lat_lon
 from uxarray.constants import INT_DTYPE, INT_FILL_VALUE
