+++ uxarray/io/_ugrid.py
@@ -1,5 +1,5 @@
 import numpy as np
-from uxarray.grid.utils import _replace_fill_values
+from uxarray.grid.connectivity import _replace_fill_values
 from uxarray.constants import INT_DTYPE, INT_FILL_VALUE
